+++ docs/changelog.md
@@ -14,6 +14,7 @@ and this project adheres to
 - `meeple import` - Import BoardGameGeek user collections.
   - Import as either one large collection or as separate collections based on item status.
   - `--dry-run` - Summarize potential import operations without persisting.
+  - `-v/--verbose` - Output additional details.
 
 ## [v0.1.0] - 2023-05-16